+++ OOP_Lab_5/src/Core/Memento/MatrixMemento.cs
@@ -8,10 +8,10 @@ namespace OOP_Lab_5.Core.Memento
 {
     public class MatrixMemento : IMatrixMemento
     {
-        public List<List<int>> Matrix { get; private set; }
+        public List<List<long>> Matrix { get; private set; }
         public int Count { get; private set; }
 
-        public MatrixMemento(List<List<int>> matrix)
+        public MatrixMemento(List<List<long>> matrix)
         {
             Matrix = matrix;
             Count = matrix.Count;
